refactor(bot): move Router, CommandHandler to command prelude
