+++ src/bot/command/prelude.rs
@@ -6,10 +6,10 @@ pub use poise::serenity_prelude::*;
 pub use crate::action_enum;
 pub use crate::action_extends;
 pub use crate::bot::checks::*;
+pub use crate::bot::command::CommandHandler;
 pub use crate::bot::command::Context;
 pub use crate::bot::command::Error;
-pub use crate::bot::controller::CommandHandler;
-pub use crate::bot::coordinator::Router;
+pub use crate::bot::command::Router;
 pub use crate::bot::error::BotError;
 pub use crate::bot::navigation::Navigation;
 pub use crate::bot::utils::*;
