+++ src/bot/mod.rs
@@ -6,8 +6,6 @@
 
 pub mod checks;
 pub mod command;
-pub mod controller;
-pub mod coordinator;
 pub mod error;
 pub mod error_handler;
 pub mod navigation;
